Proper validation of urls
